+++ docs/CHANGELOG.md
@@ -1,1 +1,1 @@
-* add Traditional Chinese translation
+* xEI recipe transfer now support for non-crafting recipes.